requirement.txt
pandas==1.4.1
joblib==1.2.0
PyYAML==6.0
tqdm==4.66.1
gdown==4.7.1
scikit-learn==1.2.0
xgboost==2.0.2
streamlit==1.29.0
requests==2.28.1
fastapi==0.108.0
numpy==1.24.4
uvicorn==0.25.0
pydantic==2.5.3